Assessment Exercise 1
The goal here was to create a program that would ask a user to input their name and an opening balance. It would then prompt a user to enter a series of transaction (wine name, price per bottle and no of bottles) and display the running balance each time. If the user inputs a -ve quantity #, they are refunded 80% of the cost. The user can enter space for a name to exit the program which should then print a receipt or statement to the computer with their details and list of transactions. This assignment was worth 10% credits. I chose to use an array to store wine object classes in each index.